naki-bot/requirements.txt
# Naki AI Mascot Bot - NakamaMesh Twitter Automation
google-genai>=1.0.0
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
